refactor: update import paths for UIWeekdaySelector component and remove obsolete file :hammer:
